M9 #146: gate commit_exec for clippy under -D warnings

+++ kernel/src/selftest/m9_linux_exec.rs
@@ -310,7 +310,7 @@ pub(crate) fn observe_syscall(frame: &SyscallContext) {
         return;
     }
 
-    if frame.rax != SYS_WRITE as u64 {
+    if frame.rax != SYS_WRITE {
         return;
     }
